Got testing to recognize the new migration folder layout, and fixed a few tests. Close but not quite everything passing just yet.

+++ tests/_support/MigrationTestMigrations/Database/Migrations/2018-01-24-102302_Another_migration.php
@@ -20,6 +20,9 @@ class Migration_another_migration extends \CodeIgniter\Database\Migration
 
 	public function down()
 	{
-		$this->forge->dropColumn('foo', 'value');
+		if ($this->db->tableExists('foo'))
+		{
+			$this->forge->dropColumn('foo', 'value');
+		}
 	}
 }
